Fixed a bug where generate_code() would crash when instructions were requested at certain times. The problem turned out to be that I relied on the order of a list of times without having sorted it. The list was obtained from the keys of a dictionary, which everyone knows come back in a seemingly random order (they're ordered by their hashes). The reason this bug wasn't picked up on sooner is that I'd used integers for the instruction times in most of my tests. A great video I watched the other day:
http://blip.tv/file/3332763

pointed out that the hash for an integer (and anything that equates to an integer with '==') *is that integer itself*.
So my keys were coming back in order most of the time, and I didn't realise til now that the list wasn't sorted. Thanks
to Russ for the bug report.