Function ramps now take relative times rather than absolute. That is, the lambdas returned by function in functions.py are evaluated at timepoints measured from the beginning of the ramp, rather than from the beginning of the experiment.
This does not affect existing labscripts, but means that new functions must be implemented accordingly.

This change was made to facilitate delayed triggering of devices -- where a device is only triggered to begin its run after t = 0. Having the lambdas take times relative to the start time of their corresponding ramps makes it easier to offset these ramps later to match the delayed trigger time.