Added a no_warnings context manager. When used, all warnings are supressed. Use it like this:
with no_warnings:
    function_that_would_normally_generate_a_warning()

This could be useful for supressing warnings when it is known that they do not pose a problem, allowing unexpected warnings to not be drowned out in the noise.

'mild' warnings (everything but value-at-certain-time-is-being-overwritten warnings) are still supressed by default, they can be re-enabled by setting config.supress_mild_warnings to False, and then supressed selectively with the context manager if desired.