gated-clocks: Fixed loss of precision for AnalogQuantities used for large numbers This affected DDS devices (frequency component only) such as the RFBlaster and NovatechDDS9m.
As an example of the bug, in the Monash Dual Species BEC lab, we effectively had a linear ramp in frequency with an fluctuation of 2.5KHz in the magnitude of the step size at each point in the ramp. This was detected because when I looked at the RFBlaster diff table, the frequency was changing by 1100, 1099, 1031, 1100, 1099, 1031.

The fix was to increase the precision of the numpy arrays to float64. Interestingly, this has not affected the size of the HDF5 files produced. This is because the data is quantised before writing the the HDF5 file. The issue affected the precision of the data that was input into the quantisation functions.