Fixed bug where the ClockLine clock_limits were not taken into account when determining the minimum time between a wait and subsequent output. This meant that triggers to secondary pseudoclocks, immediately followed by output could be faster than the allowed update rate of IntermediateDevices. This was fortunately caught further down the line, but should have never been a problem.
The wait() function (via a fix to trigger_all_pseudoclocks) now returns the correct length of time before you can command further output.

It is possible this bug only affected wait() commands with a single pseudoclock where the pseudoclock trigger_duration was very short and the wait_monitor was on a clockline that had a high clock_limit.