Added docs requirements file and better .gitignore
The docs requirements file now ensures PyQt5 is installed when building docs on readthedocs, which is necessary for sphinx autodoc to work correctly.

I've also updated the .gitignore file. This is mostly based on the generic GitHub Python .gitignore (that you can create when making a new repository) with a few extras at the end for our own rules